Moved export to model
Made export a ICommand so we have CanExecute
