+++ src/ServiceInsight/SequenceDiagram/SequenceDiagramView.xaml
@@ -30,11 +30,8 @@
 
         <Border BorderBrush="#EFEFF2" BorderThickness="0 0 0 2">
             <StackPanel Margin="5" Orientation="Horizontal">
-                <Button Click="ExportToPng">
+                <Button Command="{Binding ExportDiagramCommand}">
                     <StackPanel Orientation="Horizontal">
-                        <!--<Image Margin="0,0,5,0"
-                               VerticalAlignment="Center"
-                               Source="/Images/DiagramEndpointIcon.png" />-->
                         <TextBlock VerticalAlignment="Center" Text="Export to png" />
                     </StackPanel>
                 </Button>
